+++ CHANGES.md
@@ -8,9 +8,10 @@ None
 
 ## 0.7.0 / 2016-September-??
 
-* [TODO] Make `cambium.core/merge-logging-context!` overridable with a nesting-aware version
-* [TODO] Make `cambium.core/context-val` overridable with a nesting-aware version
-* [TODO] Drop the `cambium.nested` namespace altogether
+* Make `cambium.core/merge-logging-context!` overridable (likely with a nesting-aware version)
+* Make `cambium.core/context-val` overridable (likely with a nesting-aware version)
+* [BREAKING CHANGE - TODO] Remove parallel `cambium.core` API from `cambium.nested` namespace
+* [TODO] Move nesting-aware, add-on functionality to the `cambium.nested` namespace
 * [TODO] Validate arguments of logging calls (literals only)
 * [TODO] Consider `null` MDC attribute value as a deletion request and remove it when merging